Add name attributes to all form inputs
At this stage if we use our form we will get a error 405 Method not
Allowed.
This error will also be visable in the terminal.
This is because, by default all of flasks views handle a GET request.
So we need to tell our contact route to handle GET and POST.