Add a print(request.form["email"]) to the contact route if statement
This is to illustrate that either this, or the .get() method may
be used to access elemnts of the form data.
if the form dictionary did not have a key matching the requested data,
the .get method would return "NONE".
But using the print(request.form["email"]) if no email key is present would
result in an "EXCEPTION"